// src/types/schema.ts
export interface components {
    schemas: {
// ...
        AssociationSimple: {
            /** Format: uuid */
            id?: string;
            name?: string;
            logoUrl?: string;
            type?: components["schemas"]["EnumOrganizerType"];
        };
// ...
        EventPreview: {
            /** Format: uuid */
            id?: string;
            title?: string;
            coverImageUrl?: string;
            /** Format: date-time */
            startAt?: string;
            /** Format: date-time */
            endAt?: string;
            locationName?: string;
            status?: components["schemas"]["EnumPublicStatus"];
            organizers?: components["schemas"]["AssociationSimple"][];
            /** @description Dacă true, utilizatorul curent (din token) este deja înscris. */
            isRegistered?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}